requirements.txt
contextily==1.3.0
folium==0.14.0
geopandas==0.12.2
matplotlib==3.7.1
numpy==1.24.2
paho_mqtt==1.6.1
pyproj==3.5.0
Shapely==2.0.1
